Corrected BA bug in ecopathlite.m
In previous versions, I had dropped the BA term from the export
term, in order to be consistant with results from EwE6.  Having
confirmed that this was a bug in EwE6, I've added the term back in.
This means that results from my routine may not agree with the EwE6
results when using a model with non-zero biomass accumulation for
any groups (the results returned by my routine are the correct,
balanced ones).