Cleaned up inter-system flow calculations, fixed omnivory bug, added import
In an effort to track down a bug in my detritus EE calculations, I ended
up rewriting this section of code in a much more matrix-oriented manner.  I'm
not quite sure what I fixed (the logic behind the EwE .NET-ish way and my
matrix way is the same... the new way simply plays to Matlab's strengths), but
the EE bug was magically fixed, and the new code introduces a flow matrix that
makes it much easier to see where everything is coming from and going to.

I also added full support for import and export to and from the system (the diet
from import possibility had been overlooked in earlier versions).

Fixed a bug in the calculation of omnivory index; values now match EwE6.