Updated for prey visibility option
When I added the prey visibility option to wce/nemurokak, I missed a
few parts of this code that needed to be updated.  I'm unlikely to ever
use this option in nemurokak (doesn't really make sense for planktonic
groups), but it's easier to manage the code this way, since nemurokak
and wce share so many setup options.