New diapause scheme, and added POFe sinking rate input
I changed the diapause-related code to allow for a slightly more complex model
of copepod diapause, in both nemurokak and wce modules.  Rather than simply assigning
a single date for migration to begin for all migrating copepods, the code now uses
a start date and time span combination.  During the diapause-start timeperiod, a
user-assigned percentage of the non-migrating group is transferred to the directed-
movement group at each time step.  This allows for a more gradual migration, rather
than the all-at-once behavior I had in there previously.

Also, I added a new input parameter allowing the user to specify a POFe sinking rate
separate from the PON one.  If that parameter isn't supplied, the code falls back on
matching the sinking rate to the PON sinking rate, as in earlier versions.