Added workaround for vertical alignment issues
I addd in some code to reposition text vertically.  This will hopefully solve
some of the issues seen with tex and latex legend labels.  In my tests it works
pretty well for LaTeX text, but is hit or miss for TeX (no worse than the old
version though, so might as well try).  Thanks to Søren Enemark for the
suggested workaround.