Now skips vertical alignment fix if title added
In the previous update, I added some adjustments to the y-position of
objects to try to account for some Tex/LaTeX issues.  However, that fix
doesn't work if a title is included for the legend.  For now, just skipping
over the adjustments if title is included; will have to add the correct
adjustments when I get some time.