Changed amax to 90%, plus minor bug fix
I changed the code so that it now uses a 90% age as the upper month for
the oldest stanza group, rather than the 99% limit I had been using.  This
is just to maintain consistency with both EwE and Rpath.

Also, fixed bug where I had inconsistent use of the bab variable.