Nemuro parameter parsing and reshaping functions updated
Replaced nemuroinputparser and nemuroflxinput with nemuroParamSets.m and
nemparams2arrays.  The former reflected a lot of old syntax from the very
early development of my version of NEMURO.  That lead to some unnecessary
code and misleading function names, so I've now cleaned up the code and
provided clearer names for the functions.